refactor(chain-index): split the chain-store adapter into reading and driving
`chain_store.rs` had grown to 658 lines behind two section dividers marking
halves that answer to different callers: every RPC path reads, and only the
sync worker drives. They also fail differently — a build can fail as the
*validator's* fault, which is why `build_to` returns a source error and no read
does — so the divider was tracking a real seam.

Now `chain_store.rs` keeps the bridge that hands the store a validator, over
`chain_store/reading.rs` and `chain_store/driving.rs`. Call sites are unchanged:
both halves are re-exported from the parent, so `chain_store::block_hash` and
`chain_store::build_to` still resolve.

The split is lopsided and honestly so. Driving is two functions and 37 lines;
reading is the rest. It buys less than the adapter split did — what it buys is
that the sync worker's surface is now nameable and small, rather than being two
functions at the bottom of a file about reads.

The read tests move with the reads and lose their `adapter_tests` name, which
only existed because one file cannot hold two modules called `tests`. The
bridge keeps its own. All nine still run.
